Add an Article Search plugin
This implementation does not depend on a search engine. Instead, a private, translated field "search_data" is added to the article. Every time there is a modification to the article, this field is updated with a seachable text-representation of the article, its category, its tags, its lead_in and its placeholderfield and the plugins it contains. This allows for reasonably fast, AJAX-based search queries.