Re-release with restricted version of translation-tools
This change isn't specifically required by this package, but if NewsBlog is used on Aldryn together with another Essential Addon, the other addon will break, at this time. Aldryn is working on a fix, in the meantime, this is a tactical release to de-risk users who might wish to use more than one Essential Addon.